Rename SidebarControlsComponent to Show::ControlsComponent
Because it's no longer a sidebar
