+++ app/components/document_component.html.erb
@@ -18,7 +18,7 @@
     <div class="col-md-9">
       <%= render ExternalLinksComponent.new(document: @document) %>
       <hr>
-      <%= render SidebarControlsComponent.new(solr_document: @document, manager: helpers.can?(:manage_item, @presenter.cocina)) %>
+      <%= render Show::ControlsComponent.new(solr_document: @document, manager: helpers.can?(:manage_item, @presenter.cocina)) %>
     </div>
   </div>
