+++ app/components/show/controls_component.html.erb
